+++ playwright/tests/subjectSearch.spec.ts
@@ -3,7 +3,8 @@ import { expect } from "@playwright/test";
 import { CustomWorld } from "../support/setup";
 
 Given("I want to do a subject search", async function (this: CustomWorld) {
-  return await this.page.goto("https://digital-research-books-beta.nypl.org/");
+  // return await this.page.goto("https://digital-research-books-beta.nypl.org/");
+  return await this.page.goto(`${this.parameters.appUrl}`);
 });
 When("I change the dropdown to subject", async function (this: CustomWorld) {
   const dropdown = await this.page.$("[aria-label='Select a search category']");
@@ -23,7 +24,6 @@ Then(
   async function (this: CustomWorld) {
     const titles = this.page.locator("h2 a");
     const count = await titles.count();
-    console.log("count", count);
 
     for (let i = 0; i < 10; ++i) {
       await titles.nth(i);
@@ -34,7 +34,6 @@ Then(
         .waitFor();
       const targetlink = this.page.locator("//a[starts-with(., 'Petroleum')]");
       expect(await targetlink.isVisible()).toBe(true);
-      //await expect(this.page.locator("//dd/ul/li/a")).toHaveText("Petroleum");
       await this.page.goBack();
     }
   }