+++ next.config.js
@@ -1,9 +1,6 @@
 const path = require("path");
 
 module.exports = {
-  future: {
-    webpack5: true,
-  },
   webpack: (config) => {
     config.resolve.alias["~"] = path.resolve(__dirname);