+++ src/__tests__/componentHelpers/SearchForm.tsx
@@ -6,17 +6,17 @@ import { SearchQuery } from "~/src/types/SearchQuery";
 export const searchFormRenderTests = (query?: SearchQuery) => {
   test("Searchbar select defaults", () => {
     const expectedSearchField =
-      query && query.queries ? query.queries[0].field : inputTerms[0].key;
+      query && query.queries ? query.queries[0].field : inputTerms[0].value;
     expect(
       screen.getByRole("combobox", { name: "Select a search category" })
     ).toHaveValue(expectedSearchField);
   });
   test("Searchbar has the correct options", () => {
     const options = screen.getAllByRole("option");
-    expect(options[0]).toHaveValue(inputTerms[0].key);
-    expect(options[1]).toHaveValue(inputTerms[1].key);
-    expect(options[2]).toHaveValue(inputTerms[2].key);
-    expect(options[3]).toHaveValue(inputTerms[3].key);
+    expect(options[0]).toHaveValue(inputTerms[0].value);
+    expect(options[1]).toHaveValue(inputTerms[1].value);
+    expect(options[2]).toHaveValue(inputTerms[2].value);
+    expect(options[3]).toHaveValue(inputTerms[3].value);
   });
   test("Searchbar has correct input", () => {
     const expectedSearchValue =